+++ backend/domain/src/main/kotlin/com/example/daycare/domain/student/Student.kt
@@ -1,8 +1,8 @@
-package com.example.daycare.domain
+package com.example.daycare.domain.student
 
 data class Student(
     val id: Long,
     val name: String,
     val age: Int,
     val className: String
-) 
+)
