+++ backend/domain/src/main/kotlin/com/example/daycare/domain/student/StudentRepository.kt
@@ -1,5 +1,5 @@
-package com.example.daycare.domain
+package com.example.daycare.domain.student
 
 interface StudentRepository {
     fun findAll(): List<Student>
-} 
+}
