--- backend/infrastructure/src/main/kotlin/Main.kt
@@ -1,5 +1,0 @@
-package com.example
-
-fun main() {
-    println("Hello World!")
-}
