+++ backend/presentation/src/main/kotlin/com/example/daycare/presentation/student/ReadStudentController.kt
@@ -1,7 +1,7 @@
-package com.example.daycare.presentation
+package com.example.daycare.presentation.student
 
-import com.example.daycare.domain.Student
-import com.example.daycare.usecase.GetStudentsUseCase
+import com.example.daycare.domain.student.Student
+import com.example.daycare.usecase.student.read.ReadStudentUsecase
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.GetMapping
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/students")
 @Tag(name = "Student", description = "生徒API")
-class StudentController(private val getStudentsUseCase: GetStudentsUseCase) {
+class ReadStudentController(private val usecase: ReadStudentUsecase) {
     @GetMapping
     @Operation(summary = "生徒一覧取得")
-    fun getStudents(): List<Student> = getStudentsUseCase.execute()
-} 
+    fun getStudents(): List<Student> = usecase.execute()
+}
